Creates hook that responds to theme change - abstracts screen tracking into hook
