+++ src/utils/theme.ts
@@ -2,11 +2,3 @@ import { Appearance } from 'react-native';
 
 export const getDefaultAppearanceMode = () =>
   Appearance.getColorScheme() || 'dark';
-
-export const getPreferredAppearanceMode = () => {
-  const preferredAppearanceMode = Appearance.getColorScheme();
-  if (!preferredAppearanceMode) {
-    return 'device';
-  }
-  return preferredAppearanceMode;
-};
